### 2016-05-04 #10 @whedon
[REVIEW]: metricsgraphics: An R htmlwidget interface to the MetricsGraphics.js D3-based visualization library
**Submitting author:** @hrbrmstr (<a href="http://orcid.org/0000-0001-5670-2640">Bob Rudis</a>)
**Repository:** <a href="https://github.com/hrbrmstr/metricsgraphics" target="_blank">https://github.com/hrbrmstr/metricsgraphics</a>
**Version:** v0.9.1

- [x] **Archive:** Does the <a target="_blank" href="10.5281/zenodo.50373">software archive</a> resolve?
- [x] **Performance:** Have the performance claims of the software been confirmed?
- [x] **Functionality documentation:** Is the functionality of the software documented to a satisfactory level (e.g. API method documentation)?
