### 2016-05-04 3a6728a
Update paper.md
