### 2016-05-07 #10 @arfon
/ cc @openjournals/joss-reviewers anyone interested in reviewing this?
