### 2016-05-07 #10 @arfon
@ramnathv we basically need the reviewer to work through the checklist at the [top of this issue](https://github.com/openjournals/joss-reviews/issues/10#issue-153024367).

There are also reviewer guidelines here: http://joss.theoj.org/about#reviewer_guidelines

If you're willing to review I'll add you to the `@openjournals/joss-reviewers` GitHub team which will give you permissions to edit the checklist.

> What are the timelines involved?

As soon as you're able :smile:
