### 2016-05-07 #10 @amoeba
Hey @hrbrmstr, great package. This is the first I've sat down with the Metrics library and it's great to have an htmlwidgets interface for it. Some notes:

- License: Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?

    No. LICENSE file indicates @hrbrmstr is the sole copyright holder. Recommend adopting one of the OSI-approved LICENSES

- Version: Does the release version given match the GitHub release (v0.9.1)?

    No. The version that is installed when I run `install_packages('metricsgraphics')` is 0.9.0. Requesting guidance from @arfon on this and @hrbrmstr for comment on whether 0.9.1 has been submitted to CRAN or what its status is. I'm indifferent to this discrepancy.

- Archive: Does the software archive resolve?

    Mostly Yes: The link in this Issue doesn't work but the URL included a DOI, which I followed, which resolved to the [Zenodo page for the 0.9.1 release](http://zenodo.org/record/50373). Main README file also has the right link. Checking it off for now.

- Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

    (1,2,3) No. Recommend editing README.md to make clear.
    As an aside, repository contains a CoC (CONDUCT.MD) which I think is fantastic.

- Paper

    See this section from http://joss.theoj.org/about#reviewer_guidelines:

    > Note the paper should not include software documentation such as API (Application Programming Interface) functionality, as this should be outlined in the software documentation.

    The paper contains a light amount of software documentation. The Summary section of the paper is fine but the Core Functionality section breaks with the above recommendation. I think the Core Functionality paragraph can be removed as it is redundant with the core software documentation.

- Tests

    The test suite exists but is minimal. This is a hard package to test because it relies on a the Metrics Graphics JS library which has its own test suite. I think what the author has done is good enough.

This is a solid submission with only minor revisions needed. The code is well-written, works as advertised, has plenty of useful examples, and contains complete and helpful documentation. The paper is concise and to the point. The software has a clear research application in that it increases the accessibility of a popular JavaScript charting library by integrating it directly into R.

Grade: (2) Minor Revisions
