### 2016-05-07 #10 @wrathematics
Hope it's ok if I chime in here.  Regarding the license, their package is actually licensed MIT.  You can confirm this by checking the License line in the DESCRIPTION file.  The reason the LICENSE file contains only a few snippets is per CRAN repository rules; the idea is that you need only fill out the missing lines in the "template" [here](https://www.r-project.org/Licenses/MIT).   The rules are explained in detail [here](https://cran.r-project.org/doc/manuals/R-exts.html#Licensing).

I do think the reviewer comment makes perfect sense.  In my opinion, CRAN's rule on this is bizarre and confusing, and I happen to think it actually violates the license (though I am not a lawyer so my opinion on this doesn't matter much).  CRAN is very picky about this, however, so it's not super obvious to me how best to conform R packages to the JOSS rules.  One can distribute full licenses in `inst/`, but then someone would have to know to look for them.  One could have the full license in an additional file named COPYING, and this could be added to .Rbuildignore to avoid issues with CRAN.  But then that's not really any less confusing.

I'm thinking of submitting a few R packages myself to JOSS, so I'm very curious to see how this is ultimately resolved.
