### 2016-05-07 #10 @amoeba
Okay by me and thanks a ton for the information, @wrathematics. I'm new to R packaging and wasn't aware of any of that. Eager to see what others have to say.

From `DESCRIPTION`: License: MIT + file LICENSE

That's super confusing since 'file LICENSE' would seem to disagree with MIT (given my near-zero knowledge of copyright law). I'll have to read up on the CRAN rules. Thanks again.
