### 2016-05-07 #10 @hrbrmstr
This. Is. *Awesome*.

I'm semi-AFK for the main part of the weekend but I'll try to get on the
edits tomorrow night.

#ty to all who are carving out cycles to make JOSS work. Pls feel free to
tap me as a reviewer for other submissions. I'll be glad to chip in.

-Bob

On Sat, May 7, 2016 at 2:18 PM, Bryce Mecum <notifications@github.com>
wrote:

> Okay by me and thanks a ton for the information, @wrathematics
> <https://github.com/wrathematics>. I'm new to R packaging and wasn't
> aware of any of that. Eager to see what others have to say.
>
> From DESCRIPTION: License: MIT + file LICENSE
>
> That's super confusing since 'file LICENSE' would seem to disagree with
> MIT (given my near-zero knowledge of copyright law). I'll have to read up
> on the CRAN rules. Thanks again.
>
