### 2016-05-07 #10 @arfon
> This. Is. *Awesome*.

Yes @amoeba and @wrathematics many thanks for such a rapid and excellent review :zap:

> I'm thinking of submitting a few R packages myself to JOSS, so I'm very curious to see how this is ultimately resolved.

On the licensing issue - I would rather see a `LICENSE` file also included in the repository if possible as I think that's both more discoverable and explicit about the licensing terms.

> I'm semi-AFK for the main part of the weekend but I'll try to get on the edits tomorrow night.

@hrbrmstr - please just leave a message on this thread to let us know when you've made the updates.

> #ty to all who are carving out cycles to make JOSS work. Pls feel free to
> tap me as a reviewer for other submissions. I'll be glad to chip in.

If you're interested then perhaps you'd like to join the `@openjournals/joss-reviewers` team on GitHub. That way if any review opportunities arise then you'll be notified of them?
