### 2016-05-22 #10 @arfon
*Bump* @hrbrmstr - have you had a change to look at these suggested modifications?
