### 2016-07-12 #10 @arfon
> Bump @hrbrmstr - have you had a change to look at these suggested modifications?

👋 @hrbrmstr - are you still interested in making these edits so we can accept this submission?
