### 2016-10-08 #10 @arfon
🎏 flagging this as stale 🎏

If there's no update here by @hrbrmstr in the next month I'll assume they want this paper withdrawn.
