### 2016-12-12 #10 @arfon
> If there's no update here by @hrbrmstr in the next month I'll assume they want this paper withdrawn.

This paper is now withdrawn from JOSS.
